requirements.txt
Django>=4.0
fastapi>=0.100
uvicorn>=0.20
sqlalchemy>=2.0
python-multipart>=0.0.6
pydantic>=2.0
requests==2.31.0
python-jose==3.3.0
python-dotenv==1.0.1